impl `cursor_insert_[string/spans]`
- Moved `TextBody`, `TextSpan` & `TextAttrs` into a seperate file.
