+++ src/interface/mod.rs
@@ -20,8 +20,9 @@ pub(crate) use self::bin::UpdateContext;
 pub use self::bin::style::{
     BackImageRegion, BinStyle, BinStyleError, BinStyleErrorType, BinStyleValidation, BinStyleWarn,
     BinStyleWarnType, BinVertex, FloatWeight, ImageEffect, LineLimit, LineSpacing, Opacity,
-    TextAttrs, TextBody, TextHoriAlign, TextSpan, TextVertAlign, TextWrap, Visibility, ZIndex,
+    TextHoriAlign, TextVertAlign, TextWrap, Visibility, ZIndex,
 };
+pub use self::bin::text_body::{TextAttrs, TextBody, TextSpan};
 pub use self::bin::text_state::{PosTextCursor, TextCursor, TextCursorAffinity, TextSelection};
 pub use self::bin::{Bin, BinID, BinPostUpdate, OVDPerfMetrics};
 pub use self::color::Color;
